Feat(Request Ride) :
UC2 - Request a Ride
Actor: Passenger
Input: passenger_id (account_id), from (latitude, longitude), to (latitude, longitude)
Output: ride_id

Rules:
It should check if the account_id has is_passenger set to true.
It should check if there is no ongoing ride for the passenger with a status other than "completed." If such a ride exists, it should throw an error.
It should generate the ride_id (uuid).
It should set the status to "requested."
It should set the date to the current date.